Cleanup swift-testing migration
