+++ Sources/FoundationInternationalization/Formatting/Duration+TimeFormatStyle.swift
@@ -147,9 +147,6 @@ extension Duration {
         }
 
     }
-
-    // For testing purpose. See notes about String._Encoding
-    internal typealias _TimeFormatStyle = TimeFormatStyle
 }
 
 @available(macOS 13.0, iOS 16.0, tvOS 16.0, watchOS 9.0, *)
