+++ Tests/FoundationEssentialsTests/AttributedString/AttributedStringTestSupport.swift
@@ -10,8 +10,10 @@
 //
 //===----------------------------------------------------------------------===//
 
-#if canImport(TestSupport)
-import TestSupport
+#if canImport(FoundationEssentials)
+import FoundationEssentials
+#else
+import Foundation
 #endif
 
 #if FOUNDATION_FRAMEWORK
